Refactor home page to root page
