+++ lib/features/root/views/views.dart
@@ -1,0 +1,1 @@
+export 'root_page.dart';
