+++ lib/features/root/views/widgets/widgets.dart
@@ -1,0 +1,1 @@
+export 'root_page.dart';
